Split SvgUnit
